refactor(e2e): lift shared funding/balance helpers into PlaywrightBaseTest
GenerateRandomNsec, GetStoreWalletIdAsync, ReadAvailableBalanceSatsAsync,
PollForBalanceAsync and FundWalletViaNoteAsync were duplicated across
4-5 test files. Moved them to PlaywrightBaseTest (FundWalletViaNoteAsync
now takes the IServiceProvider explicitly since the base has no fixture
reference). Test files keep a one-line FundWalletViaNoteAsync shim that
passes _fixture.ServerTester.PayTester.ServiceProvider.

Pure refactor — 26 tests still green in ~2m, no behavior change.
