+++ NArk.E2E.Tests/PageRenderTests.cs
@@ -1,7 +1,4 @@
 using Microsoft.Playwright;
-using NBitcoin;
-using NBitcoin.DataEncoders;
-using NBitcoin.Secp256k1;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -105,18 +102,4 @@ public class PageRenderTests : PlaywrightBaseTest
             $"Expected auth challenge (302/401/403), got {resp.Status}");
     }
 
-    private static string GenerateRandomNsec()
-    {
-        Span<byte> keyBytes = stackalloc byte[32];
-        Random.Shared.NextBytes(keyBytes);
-        if (!ECPrivKey.TryCreate(keyBytes, out _))
-        {
-            keyBytes.Clear();
-            keyBytes[31] = 0x01;
-        }
-        var encoder = Encoders.Bech32("nsec");
-        encoder.StrictLength = false;
-        encoder.SquashBytes = true;
-        return encoder.EncodeData(keyBytes.ToArray(), Bech32EncodingType.BECH32);
-    }
 }
